chore: #112 - 주석 작성
