+++ BestWish/Data/Repository/AccountRepositoryImpl.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 
+/// 유저 계정 관련 레포지토리
 final class AccountRepositoryImpl: AccountRepository {
     private let manager: SupabaseOAuthManager
 
@@ -14,10 +15,12 @@ final class AccountRepositoryImpl: AccountRepository {
         self.manager = manager
     }
 
+    /// 로그아웃
     func logout() async throws {
         try await manager.signOut()
     }
 
+    /// 회원탈퇴
     func withdraw() async throws {
         try await manager.leaveService()
     }
