+++ BestWish/Presentation/Component/GroupTitleLabel.swift
@@ -7,6 +7,7 @@
 
 import UIKit
 
+/// 그룹 타이틀 라벨 (ex. 도움말, 공지사항 등)
 final class GroupTitleLabel: UILabel {
     init(title: String) {
         super.init(frame: .zero)
