+++ BestWish/Presentation/Scene/MyPage/UserInfoManagement/ViewModel/UserInfoManagementViewModel.swift
@@ -10,27 +10,34 @@ import Foundation
 import RxSwift
 import RxRelay
 
+/// 유저 정보 관리 View Model
 final class UserInfoManagementViewModel: ViewModel {
-    private let userInfoUseCase: UserInfoUseCase
-    private let accountUseCase: AccountUseCase
-    private let disposeBag = DisposeBag()
 
+    // MARK: - Action
     enum Action {
         case getAuthProvider
         case withdraw
     }
 
+    // MARK: - State
     struct State {
         let authProvider: Observable<String?>
         let error: Observable<AppError>
     }
 
-    private let _action = PublishSubject<Action>()
+    // MARK: - Internal Property
     var action: AnyObserver<Action> { _action.asObserver() }
+    let state: State
+
+    // MARK: - Private Property
+    private let _action = PublishSubject<Action>()
 
     private let _error = PublishSubject<AppError>()
     private let _authProvider = PublishSubject<String?>()
-    let state: State
+
+    private let userInfoUseCase: UserInfoUseCase
+    private let accountUseCase: AccountUseCase
+    private let disposeBag = DisposeBag()
 
     init(userInfoUseCase: UserInfoUseCase, accountUseCase: AccountUseCase) {
         self.userInfoUseCase = userInfoUseCase
@@ -54,6 +61,7 @@ final class UserInfoManagementViewModel: ViewModel {
         }.disposed(by: disposeBag)
     }
 
+    /// SNS 계정 연동 정보 불러오기
     private func getAuthProvider() {
         Task {
             do {
@@ -66,6 +74,7 @@ final class UserInfoManagementViewModel: ViewModel {
         }
     }
 
+    /// 회원 탈퇴
     private func withdraw() {
         Task {
             do {
@@ -76,6 +85,7 @@ final class UserInfoManagementViewModel: ViewModel {
         }
     }
 
+    /// 에러 핸들링
     private func handleError(_ error: Error) {
         if let error = error as? AppError {
             _error.onNext(error)
